Organize the UI code
